requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
urllib3>=2.0.0
webdriver-manager>=4.0.0
tensorflow>=2.13.0
Werkzeug>=2.3.0